+++ server/middlewares/validators/index.js
@@ -3,6 +3,7 @@ const createTown = require('./createTown');
 const createUser = require('./createUser');
 const editTown = require('./editTown');
 const addShantytownActor = require('./shantytownActors/addShantytownActor');
+const updateShantytownActor = require('./shantytownActors/updateShantytownActor');
 
 module.exports = {
     createContact,
@@ -11,5 +12,6 @@ module.exports = {
     createUser,
     shantytownActors: {
         addShantytownActor,
+        updateShantytownActor,
     },
 };
